Fill an Array
A program that will ask the user for a number of digits to be stored. The program then creates an array of that size and then prompts the user for numbers to fill each position in the array. When all of the positions are filled, it displays the contents of the array to the user.